requirements.txt
Django
Pillow
pytesseract
PyMuPDF
djangorestframework